src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=src/requirements.txt src/requirements/requirements.in
#
amqp==5.2.0
    # via kombu
asgiref==3.7.2
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
    #   django-cors-headers
    #   django-countries
    #   django-htmx
asttokens==2.4.1
    # via icecream
attrs==23.2.0
    # via
    #   automat
    #   service-identity
    #   twisted
autobahn==23.6.2
    # via daphne
automat==22.10.0
    # via twisted
autopep8==2.0.4
    # via -r src/requirements/requirements.in
billiard==4.2.0
    # via celery
boto3==1.34.34
    # via -r src/requirements/requirements.in
botocore==1.34.34
    # via
    #   boto3
    #   s3transfer
cachetools==5.3.2
    # via -r src/requirements/requirements.in
celery==5.3.6
    # via
    #   -r src/requirements/requirements.in
    #   django-celery-beat
celery-progress==0.3
    # via -r src/requirements/requirements.in
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via cryptography
channels==4.0.0
    # via channels-redis
channels-redis==4.2.0
    # via -r src/requirements/requirements.in
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   icecream
    #   loguru
    #   qrcode
constantly==23.10.4
    # via twisted
cron-descriptor==1.4.3
    # via django-celery-beat
cryptography==42.0.2
    # via
    #   -r src/requirements/requirements.in
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==4.0.0
    # via -r src/requirements/requirements.in
dj-database-url==2.1.0
    # via -r src/requirements/requirements.in
django==5.0.6
    # via
    #   -r src/requirements/requirements.in
    #   channels
    #   dj-database-url
    #   django-celery-beat
    #   django-components
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-htmx
    #   django-js-asset
    #   django-storages
    #   django-subdomains
    #   django-timezone-field
    #   djangorestframework
django-celery-beat==2.5.0
    # via -r src/requirements/requirements.in
django-components==0.67
    # via -r src/requirements/requirements.in
django-cors-headers==4.3.1
    # via -r src/requirements/requirements.in
django-countries==7.5.1
    # via -r src/requirements/requirements.in
django-extensions==3.2.3
    # via -r src/requirements/requirements.in
django-filter==23.5
    # via -r src/requirements/requirements.in
django-htmx==1.17.3
    # via -r src/requirements/requirements.in
django-js-asset==2.2.0
    # via django-mptt
django-mptt==0.16.0
    # via -r src/requirements/requirements.in
django-split-settings==1.2.0
    # via -r src/requirements/requirements.in
django-storages==1.14.2
    # via -r src/requirements/requirements.in
django-subdomains==2.1.0
    # via -r src/requirements/requirements.in
django-timezone-field==6.1.0
    # via django-celery-beat
django-widget-tweaks==1.5.0
    # via -r src/requirements/requirements.in
djangorestframework==3.14.0
    # via -r src/requirements/requirements.in
executing==2.0.1
    # via icecream
genshi==0.7.7
    # via relatorio
gunicorn==21.2.0
    # via -r src/requirements/requirements.in
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
icecream==2.1.3
    # via -r src/requirements/requirements.in
idna==3.6
    # via
    #   hyperlink
    #   requests
    #   twisted
incremental==22.10.0
    # via twisted
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.3.5
    # via celery
loguru==0.7.2
    # via -r src/requirements/requirements.in
lxml==5.1.0
    # via relatorio
msgpack==1.0.7
    # via channels-redis
mswinprint==1.1
    # via -r src/requirements/requirements.in
numpy==1.26.3
    # via pandas
packaging==23.2
    # via gunicorn
pandas==2.2.0
    # via -r src/requirements/requirements.in
pillow==10.2.0
    # via -r src/requirements/requirements.in
prompt-toolkit==3.0.43
    # via click-repl
psycopg2==2.9.9
    # via -r src/requirements/requirements.in
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.3.0
    # via service-identity
pycodestyle==2.11.1
    # via autopep8
pycparser==2.21
    # via cffi
pycryptodome==3.20.0
    # via -r src/requirements/requirements.in
pygments==2.17.2
    # via icecream
pyopenssl==24.0.0
    # via twisted
pypng==0.20220715.0
    # via qrcode
python-crontab==3.0.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   botocore
    #   celery
    #   pandas
    #   python-crontab
python-decouple==3.8
    # via -r src/requirements/requirements.in
pytz==2024.1
    # via
    #   -r src/requirements/requirements.in
    #   djangorestframework
    #   pandas
pywin32==306
    # via
    #   -r src/requirements/requirements.in
    #   wmi
qrcode==7.4.2
    # via -r src/requirements/requirements.in
redis==5.0.1
    # via channels-redis
relatorio==0.10.1
    # via -r src/requirements/requirements.in
requests==2.31.0
    # via stripe
s3transfer==0.10.0
    # via boto3
service-identity==24.1.0
    # via twisted
six==1.16.0
    # via
    #   -r src/requirements/requirements.in
    #   asttokens
    #   automat
    #   genshi
    #   python-dateutil
sqlparse==0.4.4
    # via django
stripe==8.1.0
    # via -r src/requirements/requirements.in
twisted[tls]==23.10.0
    # via daphne
twisted-iocpsupport==1.0.4
    # via twisted
txaio==23.1.1
    # via autobahn
typing-extensions==4.9.0
    # via
    #   dj-database-url
    #   django-countries
    #   qrcode
    #   stripe
    #   twisted
tzdata==2023.4
    # via
    #   celery
    #   django
    #   django-celery-beat
    #   pandas
    #   tzlocal
tzlocal==5.2
    # via -r src/requirements/requirements.in
urllib3==2.0.7
    # via
    #   botocore
    #   requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
websockets==12.0
    # via -r src/requirements/requirements.in
win32-setctime==1.1.0
    # via loguru
wmi==1.5.1
    # via -r src/requirements/requirements.in
zope-interface==6.1
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
